Move iterable methods from array to iterable traits
