+++ src/Type/JustNullableTypeTrait.php
@@ -4,12 +4,14 @@ namespace PHPStan\Type;
 
 use PHPStan\TrinaryLogic;
 use PHPStan\Type\Traits\NonArrayTypeTrait;
+use PHPStan\Type\Traits\NonIterableTypeTrait;
 use function get_class;
 
 trait JustNullableTypeTrait
 {
 
 	use NonArrayTypeTrait;
+	use NonIterableTypeTrait;
 
 	/**
 	 * @return string[]
