+++ src/Type/NeverType.php
@@ -15,6 +15,7 @@ use PHPStan\TrinaryLogic;
 use PHPStan\Type\Traits\NonArrayTypeTrait;
 use PHPStan\Type\Traits\NonGeneralizableTypeTrait;
 use PHPStan\Type\Traits\NonGenericTypeTrait;
+use PHPStan\Type\Traits\NonIterableTypeTrait;
 use PHPStan\Type\Traits\NonRemoveableTypeTrait;
 use PHPStan\Type\Traits\UndecidedBooleanTypeTrait;
 use PHPStan\Type\Traits\UndecidedComparisonCompoundTypeTrait;
@@ -26,6 +27,7 @@ class NeverType implements CompoundType
 	use UndecidedBooleanTypeTrait;
 	use NonArrayTypeTrait;
 	use NonGenericTypeTrait;
+	use NonIterableTypeTrait;
 	use UndecidedComparisonCompoundTypeTrait;
 	use NonRemoveableTypeTrait;
 	use NonGeneralizableTypeTrait;
