+++ src/Type/Traits/MaybeIterableTypeTrait.php
@@ -24,7 +24,27 @@ trait MaybeIterableTypeTrait
 		return new MixedType();
 	}
 
+	public function getFirstIterableKeyType(): Type
+	{
+		return new MixedType();
+	}
+
+	public function getLastIterableKeyType(): Type
+	{
+		return new MixedType();
+	}
+
 	public function getIterableValueType(): Type
+	{
+		return new MixedType();
+	}
+
+	public function getFirstIterableValueType(): Type
+	{
+		return new MixedType();
+	}
+
+	public function getLastIterableValueType(): Type
 	{
 		return new MixedType();
 	}
