+++ src/Type/Traits/NonArrayTypeTrait.php
@@ -3,8 +3,6 @@
 namespace PHPStan\Type\Traits;
 
 use PHPStan\TrinaryLogic;
-use PHPStan\Type\ErrorType;
-use PHPStan\Type\Type;
 
 trait NonArrayTypeTrait
 {
@@ -17,26 +15,6 @@ trait NonArrayTypeTrait
 	public function getConstantArrays(): array
 	{
 		return [];
-	}
-
-	public function getFirstIterableKeyType(): Type
-	{
-		return new ErrorType();
-	}
-
-	public function getLastIterableKeyType(): Type
-	{
-		return new ErrorType();
-	}
-
-	public function getFirstIterableValueType(): Type
-	{
-		return new ErrorType();
-	}
-
-	public function getLastIterableValueType(): Type
-	{
-		return new ErrorType();
 	}
 
 	public function isArray(): TrinaryLogic
