+++ src/Type/Traits/NonIterableTypeTrait.php
@@ -24,7 +24,27 @@ trait NonIterableTypeTrait
 		return new ErrorType();
 	}
 
+	public function getFirstIterableKeyType(): Type
+	{
+		return new ErrorType();
+	}
+
+	public function getLastIterableKeyType(): Type
+	{
+		return new ErrorType();
+	}
+
 	public function getIterableValueType(): Type
+	{
+		return new ErrorType();
+	}
+
+	public function getFirstIterableValueType(): Type
+	{
+		return new ErrorType();
+	}
+
+	public function getLastIterableValueType(): Type
 	{
 		return new ErrorType();
 	}
